Adds support for terminating multiple pods per run
The configuration flag --max-kill specifies the maximum number of pods to be terminated per run,
with a default value of 1, to maintain existing behavior.

Adds a new filter: filterByOwnerReference to avoid terminating multiple pods belonging to the
same owner. This prevents pods deployed together to be terminated at the same time.

In practice, no more than 1 pod per owner will be terminated during each run. All pods without an owner
reference are eligible to be terminated in the same batch (up to --max-kill value)