Improve Helm experience
Since the deprecation of the `stable` Helm chart repository by Helm, chaoskube's chart is one of the charts I've been waiting for to be ported.

Initially, when that happened I didn't get it to work and didn't look too deep into it. Now that I'm building a home cluster it came up again,
and after looking into it properly there are a few things that came up:

- [X] The project's readme still lists the old `stable` Helm chart repository.
- [X] The chart's readme lists a non-existent (?) Helm chart repository
- [X] Has default values for arguments causing issues for some of those when empty
- [X] Missing `v` for the default tag

P.S. This PR only aims to make it work out of the box again and contains some breaking changes causing the major version bump.